fixing negative sum

+++ cp1/sum_direct.cpp
@@ -20,7 +20,7 @@ setup(int64_t N, uint64_t A[])
 int64_t
 sum(int64_t N, uint64_t A[])
 {
-   int sum  = 0;
+   int64_t sum  = 0;
    for(int i =0; i<N; i++){
       sum += A[i];
    }
